+++ nagra/schema.py
@@ -7,6 +7,7 @@ from typing import Optional, TYPE_CHECKING
 from warnings import warn
 
 import toml
+from nagra.exceptions import IncorrectSchema
 from nagra.statement import Statement
 from nagra.transaction import DummyTransaction, Transaction
 from nagra.utils import logger, snake_to_pascal, template
@@ -293,11 +294,18 @@ class Schema:
                 warn(MSSQL_ARRAY_MSG.format(table=table.name), RuntimeWarning)
                 continue
 
+            if table.nullable and trn.flavor != "postgresql":
+                raise IncorrectSchema(
+                    f"Table '{table.name}': nullable natural key columns are only supported for postgresql"
+                )
+
+            # Columns that are neither primary key nor foreign keys.
             columns_not_pk_fk = [
                 col
                 for col in table.columns
                 if col != table.primary_key and col not in table.foreign_keys
             ]
+            required = filter(lambda c: table.required(c), columns_not_pk_fk)
 
             pk_fk_table = (
                 self.tables.get(table.foreign_keys[table.primary_key])
@@ -314,7 +322,7 @@ class Schema:
                 table=table,
                 columns=columns_not_pk_fk,
                 ctypes=ctypes,
-                not_null=table.not_null,
+                not_null=required,
                 default=table.default,
                 pk_fk_table=pk_fk_table,
                 natural_key=table.natural_key,
@@ -387,12 +395,18 @@ class Schema:
                 continue
             if not table.natural_key:
                 continue
+            kwargs = {}
+            if trn.flavor == "postgresql":
+                kwargs["nulls_not_distinct"] = any(
+                    not table.required(col) for col in table.natural_key
+                )
 
             stmt = Statement(
                 "create_unique_index",
                 trn.flavor,
                 table=name,
                 natural_key=table.natural_key,
+                **kwargs,
             )
             yield stmt()