+++ nagra/template/mssql/create_unique_index.sql
@@ -3,5 +3,6 @@ CREATE UNIQUE INDEX [{{ table ~ "_idx" }}] ON [{{ table }}] (
   {%- for col in natural_key -%}
   [{{ col }}]{{ ", " if not loop.last }}
   {%- endfor %}
-);
+)
+;
 {%- endif %}