+++ nagra/template/sqlite/create_unique_index.sql
@@ -1,5 +1,6 @@
 {%- if natural_key is defined -%}
 CREATE UNIQUE INDEX {{table}}_idx ON "{{table}}" (
   {{ natural_key | map('autoquote') |join(', ') }}
-);
+)
+;
 {%- endif %}